* src/minikafs.c(minikafs_5log): initialize use_ccache as a handle for the 	default cred cache, not ccache, when ccache is NULL. * src/options.c(option_t): add, for parsing a value as a krb5_deltat if it 	can't be parsed as a normal integer. * src/options.h: change normal and renewable lifetimes to krb5_deltat * src/options.c(_pam_krb5_options_init): parse lifetimes using option_t 	instead of option_i. * src/*.c: random signed/unsigned warning corrections.

+++ src/prompter.c
@@ -138,7 +138,7 @@ _pam_krb5_prompter(krb5_context context, void *data,
 			_pam_krb5_maybe_free_responses(responses, num_prompts);
 			return KRB5_LIBOS_CANTREADPWD;
 		}
-		if (xstrlen(responses[i + offset].resp) >= prompts[i].reply->length) {
+		if ((unsigned int)xstrlen(responses[i + offset].resp) >= prompts[i].reply->length) {
 			_pam_krb5_maybe_free_responses(responses, num_prompts);
 			return KRB5_LIBOS_CANTREADPWD;
 		}
